requirements.txt
pytest>=7.0
gitpython>=3.1.0
sentence-transformers>=2.2.0
torch>=2.0.0
numpy>=1.24.0,<2.0.0
mcp>=0.9.0
tree-sitter>=0.20.0
tree-sitter-python>=0.20.0
tree-sitter-javascript>=0.20.0
tree-sitter-typescript>=0.20.0
tree-sitter-c-sharp>=0.20.0
pyyaml>=6.0
